requirements.txt
alembic==1.4.1
Babel==2.8.0
blinker==1.4
click==7.1.1
coverage==5.0.3
decorator==4.4.2
defusedxml==0.6.0
dominate==2.5.1
Flask==1.1.1
Flask-Babel==1.0.0
Flask-Bootstrap==3.3.7.1
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==2.5.2
Flask-OpenID==1.2.5
Flask-Scss==0.5
Flask-Sijax==0.4.1
Flask-SQLAlchemy==2.4.1
Flask-WhooshAlchemy==0.56
Flask-WTF==0.14.3
flipflop==1.0
future==0.18.2
guess-language==0.2
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.11.1
libsass==0.19.4
Mako==1.1.2
MarkupSafe==1.1.1
pbr==5.4.4
psycopg2==2.8.4
psycopg2-binary==2.8.4
pyScss==1.3.7
python-dateutil==2.8.1
python-editor==1.0.4
python3-openid==3.1.0
pytz==2019.3
Sijax==0.3.2
six==1.14.0
SQLAlchemy==1.3.13
sqlalchemy-migrate==0.13.0
sqlparse==0.3.1
Tempita==0.5.2
visitor==0.1.3
Werkzeug==1.0.0
Whoosh==2.7.4
WTForms==2.2.1
